correcao do resultado da divisao na calculadora

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -17,7 +17,7 @@ public class Calculadora {
 		int soma = soma(a, b);
 		int subtracao = subtracao(a, b);
 		int multiplicacao = multiplicacao(a, b);
-		int divisao = divisao(a, b);
+		double divisao = divisao(a, b);
 		
 		System.out.println("soma: " + soma);
 		System.out.println("sub: " + subtracao);
@@ -33,8 +33,8 @@ public class Calculadora {
 		return a - b;
 	}
 	
-	public static int divisao(int a, int b) {
-		return a / b;
+	public static double divisao(int a, int b) {
+		return (double) a / b;
 	}
 	
 	public static int multiplicacao(int a, int b) {
